+++ Examples/AutoCoding/Classes/TodoList.m
@@ -24,16 +24,13 @@
 
 - (void)setUp
 {
-	items = [[NSMutableArray alloc] init];
+    self.items = [NSMutableArray array];
 }
 
 - (void)setWithArray:(NSArray *)array
 {
 	//initialise with default list from plist
-	for (NSString *label in array)
-	{
-		[items addObject:[TodoItem instanceWithLabel:label]];
-	}
+	[items setArray:[TodoItem instancesWithArray:array]];
 }
 
 //note: we've not implemented the NSCoding methods